+++ src/ragtime/base/llm_class.py
@@ -1,0 +1,138 @@
+#!/usr/bin/env python3
+
+from abc import abstractmethod
+
+from ragtime.base.data_type import *
+from ragtime.base.prompter import *
+
+from ragtime.expe import (
+    LLMAnswer,
+    RagtimeBase,
+    QA,
+    Prompt,
+    WithLLMAnswer
+)
+from ragtime.config import RagtimeException, logger
+
+from litellm import ( completion_cost, acompletion )
+from datetime import ( datetime )
+from typing import ( Optional )
+import asyncio
+
+class LLM(RagtimeBase):
+    """
+    Base class for text to text LLMs.
+    Class deriving from LLM must implement `complete`.
+    A Prompter must be provided at creation time.
+    Instantiates a get_prompt so as to be able change the prompt LLM-wise.
+    """
+    name:Optional[str] = None
+    prompter:Prompter
+
+    async def generate(
+            self,
+            cur_obj:WithLLMAnswer,
+            prev_obj:WithLLMAnswer,
+            qa:QA,
+            start_from:StartFrom,
+            b_missing_only:bool,
+            **kwargs
+    ) -> WithLLMAnswer:
+        """
+        Generate prompt and execute LLM
+        Returns the retrieved or created object containing the LLMAnswer
+        If None, LLMAnswer retrieval or generation went wrong and post-processing
+        must be skipped
+        """
+
+        assert not prev_obj or (cur_obj.__class__ == prev_obj.__class__)
+        cur_class_name:str = cur_obj.__class__.__name__
+
+        # Get prompt
+        if not(prev_obj and prev_obj.llm_answer and prev_obj.llm_answer.prompt) or \
+        (start_from <= StartFrom.prompt and not b_missing_only):
+            logger.debug(f'Either no {cur_class_name} / LLMAnswer / Prompt exists yet, or you asked to regenerate Prompt ==> generate prompt')
+            prompt = self.prompter.get_prompt(**kwargs)
+        else:
+            logger.debug(f'Reuse existing Prompt')
+            prompt = prev_obj.llm_answer.prompt
+
+        # Generates text
+        result:WithLLMAnswer = cur_obj
+        if not(prev_obj and prev_obj.llm_answer) or (start_from <= StartFrom.llm and not b_missing_only):
+            logger.debug(f'Either no {cur_class_name} / LLMAnswer exists yet, or you asked to regenerate it ==> generate LLMAnswer')
+            try:
+                result.llm_answer = await self.complete(prompt)
+            except Exception as e:
+                logger.exception(f'Exception while generating - skip it\n{e}')
+                result = None
+        else:
+            logger.debug(f'Reuse existing LLMAnswer in {cur_class_name}')
+            result = prev_obj
+
+        # Post-process
+        if result.llm_answer and (not (prev_obj and prev_obj.llm_answer) or not b_missing_only and start_from <= StartFrom.post_process):
+            logger.debug(f'Post-process {cur_class_name}')
+            self.prompter.post_process(qa=qa, cur_obj=result)
+        else:
+            logger.debug('Reuse post-processing')
+
+        return result
+
+    @abstractmethod
+    async def complete(self, prompt:Prompt) -> LLMAnswer:
+        raise NotImplementedError('Must implement this!')
+
+class LiteLLM(LLM):
+    """
+    Simple extension of LLM based on the litellm library.
+    Allows to call LLMs by their name in a stantardized way.
+    The default get_prompt method is not changed.
+    The generate method uses the standard litellm completion method.
+    Default values of temperature (0.0)
+    Number of retries when calling the API (3) can be changed.
+    The proper API keys and endpoints have to be specified in the keys.py module.
+    """
+    name:str
+    temperature:float = 0.0
+    num_retries:int = 3
+
+    async def complete(self, prompt:Prompt) -> LLMAnswer:
+        messages:list[dict] = [{"content":prompt.system, "role":"system"},
+                               {"content":prompt.user, "role":"user"}]
+        retry:int = 1
+        wait_step:float = 3.0
+        start_ts:datetime = datetime.now()
+        answer:dict = None
+        while retry < self.num_retries:
+            try:
+                time_to_wait:float = wait_step
+                answer = await acompletion(
+                    messages = messages,
+                    model = self.name,
+                    temperature = self.temperature,
+                    num_retries = self.num_retries,
+                    max_tokens = self.max_tokens,
+                )
+                retry = 0
+            except Exception as e:
+                logger.debug(f'Rate limit reached - will retry in {time_to_wait:.2f}s ({str(e)})')
+                await asyncio.sleep(time_to_wait)
+                retry += 1
+            except Exception as e:
+                logger.exception(f'The following exception occurred with prompt {prompt}' + '\n' + str(e))
+                return None
+
+        full_name:str = answer['model']
+        text:str = answer['choices'][0]['message']['content']
+        duration:float = answer._response_ms/1000 if hasattr(answer, "_response_ms") else None # sometimes _response_ms is not present
+        cost:float = float(completion_cost(ans))
+        return LLMAnswer(
+            name = self.name,
+            full_name= full_name,
+            prompt = prompt,
+            text = text,
+            timestamp = start_ts,
+            duration = duration,
+            cost = cost
+        )
