+++ src/ragtime/config.toml
@@ -1,0 +1,47 @@
+
+[Configuration]
+envFile =
+
+llms_name = [
+    "gpt-4",
+    'vertex_ai/gemini-pro',
+    "mistral/mistral-large-latest",
+    "groq/llama3-8b-8192",
+    "groq/llama3-70b-8192",
+    "groq/mixtral-8x7b-32768",
+    "groq/gemma-7b-it"
+]
+prompter = ''
+
+[Configuration.file.input]
+path =
+folder =
+name =
+
+[Configuration.file.output]
+path =
+folder =
+name =
+
+[Configuration.api_keys]
+OPENAI_API_KEY =
+ALEPHALPHA_API_KEY =
+ANTHROPIC_API_KEY =
+COHERE_API_KEY =
+HUGGINGFACE_API_KEY =
+MISTRAL_API_KEY =
+NLP_CLOUD_API_KEY =
+GROQ_API_KEY =
+
+
+[Generate.questions]
+[Generate.answers]
+[Generate.facts]
+[Generate.evals]
+
+[Export.html]
+[Export.html.file.template]
+[Export.html.file.input]
+[Export.html.file.output]
+
+[Export.spreadsheet]
