+++ src/ragtime/expe.py
@@ -10,7 +10,20 @@ from copy import copy
 from datetime import datetime
 from typing import Any, Callable, Generic, Optional, TypeVar, Union
 from pydantic import BaseModel, Field
-from ragtime.config import DEFAULT_FACTS_COL, DEFAULT_HUMAN_EVAL_COL, DEFAULT_HEADER_SIZE, DEFAULT_HUMAN_EVAL_COL, DEFAULT_ANSWERS_COL, DEFAULT_QUESTION_COL, DEFAULT_SPREADSHEET_TEMPLATE, DEFAULT_WORKSHEET, RagtimeException, logger, DEFAULT_HTML_RENDERING, DEFAULT_HTML_TEMPLATE
+from ragtime.config import (
+    DEFAULT_FACTS_COL,
+    DEFAULT_HUMAN_EVAL_COL,
+    DEFAULT_HEADER_SIZE,
+    DEFAULT_HUMAN_EVAL_COL,
+    DEFAULT_ANSWERS_COL,
+    DEFAULT_QUESTION_COL,
+    DEFAULT_SPREADSHEET_TEMPLATE,
+    DEFAULT_WORKSHEET,
+    RagtimeException,
+    logger,
+    DEFAULT_HTML_RENDERING,
+    DEFAULT_HTML_TEMPLATE
+    )
 from jinja2 import Environment, FileSystemLoader
 from tabulate import tabulate
