+++ src/ragtime/prompters.py
@@ -1,0 +1,23 @@
+#!/usr/bin/env python3
+
+from ragtime.base.data_type import *
+
+from ragtime.generator.prompters.base_answers import *
+from ragtime.generator.prompters.eval_french import *
+from ragtime.generator.prompters.fact_french import *
+from ragtime.generator.prompters.RAG_answer_french import *
+from ragtime.generator.prompters.simple_eval_french import *
+from ragtime.generator.prompters.simple_fact_french import *
+from ragtime.generator.prompters.two_fact_eval_french import *
+
+"""
+PptrBaseAns
+PptrRAGAnsFR
+
+PptrSimpleFactsFR
+PptrFactsFRv2
+
+PptrSimpleEvalFR
+PptrEvalFRv2
+PptrTwoFactsEvalFR
+"""
